+++ solver/src/main/java/org/chocosolver/solver/search/strategy/selectors/variables/AbstractCriterionBasedVariableSelector.java
@@ -20,7 +20,6 @@ import org.chocosolver.solver.exception.ContradictionException;
 import org.chocosolver.solver.search.loop.monitors.IMonitorContradiction;
 import org.chocosolver.solver.search.loop.monitors.IMonitorRestart;
 import org.chocosolver.solver.variables.IVariableMonitor;
-import org.chocosolver.solver.variables.IntVar;
 import org.chocosolver.solver.variables.Variable;
 import org.chocosolver.solver.variables.events.IEventType;
 import org.chocosolver.solver.variables.events.IntEventType;
@@ -39,8 +38,8 @@ import java.util.stream.Collectors;
  * @author Charles Prud'homme
  * @since 26/02/2020.
  */
-public abstract class AbstractCriterionBasedVariableSelector implements VariableSelector<IntVar>,
-        IVariableMonitor<Variable>, IMonitorContradiction, IMonitorRestart {
+public abstract class AbstractCriterionBasedVariableSelector<V extends Variable> implements VariableSelector<V>,
+        IVariableMonitor<V>, IMonitorContradiction, IMonitorRestart {
 
     /**
      * An element helps to keep 2 things up to date:
@@ -120,7 +119,7 @@ public abstract class AbstractCriterionBasedVariableSelector implements Variable
     final HashMap<Propagator<?>, double[]> refinedWeights = new HashMap<>();
     static final double[] rw = {0.};
 
-    public AbstractCriterionBasedVariableSelector(IntVar[] vars, long seed, int flush) {
+    public AbstractCriterionBasedVariableSelector(V[] vars, long seed, int flush) {
         this.random = new java.util.Random(seed);
         this.solver = vars[0].getModel().getSolver();
         this.environment = vars[0].getModel().getEnvironment();
@@ -129,8 +128,8 @@ public abstract class AbstractCriterionBasedVariableSelector implements Variable
     }
 
     @Override
-    public final IntVar getVariable(IntVar[] vars) {
-        IntVar best = null;
+    public final V getVariable(V[] vars) {
+        V best = null;
         bests.resetQuick();
         double w = Double.NEGATIVE_INFINITY;
         int to = last.get();
@@ -148,7 +147,7 @@ public abstract class AbstractCriterionBasedVariableSelector implements Variable
                 }
             } else {
                 // swap
-                IntVar tmp = vars[to];
+                V tmp = vars[to];
                 vars[to] = vars[idx];
                 vars[idx] = tmp;
                 idx--;
@@ -164,7 +163,7 @@ public abstract class AbstractCriterionBasedVariableSelector implements Variable
         return best;
     }
 
-    protected abstract double weight(IntVar v);
+    protected abstract double weight(V v);
 
     @Override
     public final void onContradiction(ContradictionException cex) {